+++ src/app/(usercard)/my-activities/page.tsx
@@ -1,0 +1,7 @@
+const Page = () => {
+  return (
+    <div>액티비티</div>
+  )
+}
+
+export default Page
